pyproject.toml
[tool.poetry]
name = "sup"
version = "0.1.0"
description = ""
authors = ["Zev Averbach <zev@averba.ch>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
rich = "^13.7.0"
typer = "^0.9.0"
toml = "^0.10.2"
sqlmodel = "^0.0.14"
mysql-connector-python = "^8.3.0"
python-dotenv = "^1.0.1"


[tool.poetry.group.dev.dependencies]
pylint = "^3.0.3"
ruff = "^0.2.0"
black = "^24.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
sup = "sup.cli:app"
